refactor: move queryplanmetadata into pkg/query and wire it into the DispatchExecutor
